+++ src/types.ts
@@ -1,6 +1,7 @@
 export type Theme = 'dark' | 'light' | string;
 export type Breakpoint = 'xs' | 'sm' | 'md' | 'lg' | 'xl' | '2xl' | string;
 export type State = 'hover' | 'active' | 'focus' | 'visited' | 'focus-visible' | 'focus-within' | 'disabled' | 'enabled' | 'checked' | string;
+export type BreakpointStrategy = 'mobile-first' | 'desktop-first' | 'exact';
 
 export interface StyleConditions {
   theme?: Theme;
@@ -22,4 +23,5 @@ export interface GetStyleOptions {
   theme?: Theme;
   state?: State;
   breakpoint?: Breakpoint;
+  breakpointStrategy?: BreakpointStrategy;
 }